backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
pymongo==4.6.1
python-dotenv==1.0.0
bcrypt==4.1.2
google-cloud-storage==2.14.0
azure-storage-blob==12.19.0
Werkzeug==3.0.1
email-validator==2.1.0
certifi==2024.2.2
dnspython==2.5.0
